requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.7.0
sqlalchemy>=2.0.0
python-dotenv
openai>=1.40.0
pdfplumber
python-docx